Move PostgreSQLPreparedStatementRegistry into ConnectionSession
